backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymongo==4.10.1
python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
email-validator==2.2.0

